{{cookiecutter.project_slug}}/requirements/local.txt
# Local development dependencies go here

-r base.txt

Sphinx==1.4.6
factory_boy==2.7.0

# improved REPL
ipdb==0.10.1

pytest==3.0.1
pytest-tornado==0.4.5
pytest-sugar==0.7.1
pytest-cov==2.3.1
